Moved DenomClass method to its own file
